moved apply_parameters
